% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCallGeno.R
\name{batchCallGeno}
\alias{batchCallGeno}
\title{Genotyping all sample by chromsome}
\usage{
batchCallGeno(
  inputfile,
  datatype,
  outputfile,
  father = "1/1",
  mother = "0/0",
  screening = FALSE,
  maf = NA,
  geno = NA,
  mind = NA,
  hwe = NA,
  window.type = "number",
  window.size = 15,
  low = 0.2,
  high = 0.8,
  fix = TRUE,
  fix.size = 10,
  filetype = NA
)
}
\arguments{
\item{inputfile}{input file name}

\item{datatype}{inputfile type. support vcf/ped/bed}

\item{outputfile}{output file name}

\item{father}{support a character vector or a number that father index of sample}

\item{mother}{support a character vector or a number that father index of sample}

\item{screening}{a logical value. use plink to screening}

\item{maf}{if screening == T ,it useful}

\item{geno}{if screening == T ,it useful}

\item{mind}{if screening == T ,it useful}

\item{hwe}{if screening == T ,it useful}

\item{window.type}{"number" ,"kb" two types of call window default is number}

\item{window.size}{window size dafalut is 15}

\item{low}{default is 0.2}

\item{high}{default is 0.8}

\item{fix}{a logical value ,default is T}

\item{fix.size}{define the neighbor size to fix the error}

\item{filetype}{output file type}
}
\value{
files of relust
}
\description{
Genotyping all sample by chromsome
}
