% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMap-package.R
\docType{package}
\name{BinMap-package}
\alias{BinMap}
\alias{BinMap-package}
\title{BinMap: Call Marker from SNP Data using Binmap}
\description{
The raw NGS (Next Generation Sequencing) variants called 
    from GBS (Genotyping by Sequencing) / WES (Whole Exon Sequencing)/
    WGS (Whole Genome Sequencing) may include many error sites. The 
    'binmapr' could fix the potential error sites and generate highly
    confident markers for downstream analysis, such as QTL (quantitative
    trait locus) mapping, genetic map construction. 
    Davey, J.W. (2011) <doi:10.1038/nrg3012>.
}
\keyword{internal}
