% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callWindowGeno.R
\name{callWindowGeno}
\alias{callWindowGeno}
\title{Call genotype}
\usage{
callWindowGeno(
  x,
  window.type = "number",
  window.size = 15,
  low = 0.2,
  high = 0.8
)
}
\arguments{
\item{x}{data.frame}

\item{window.type}{"number" ,"kb" two types of call window default is number}

\item{window.size}{window size dafalut is 15}

\item{low}{default is 0.2}

\item{high}{default is 0.8}
}
\value{
data.frame
}
\description{
get genotype of a chrom of single sample
}
