% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixGeno.R
\name{fixGeno}
\alias{fixGeno}
\title{Fix potential genotype error}
\usage{
fixGeno(x, fix.size = 10)
}
\arguments{
\item{x}{a vector}

\item{fix.size}{define the neighbor size to fix the error}
}
\value{
a vector
}
\description{
fix potential genotype error
}
