% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfiletype.R
\name{outfiletype}
\alias{outfiletype}
\title{Choose output file type}
\usage{
outfiletype(x, outputfile, filetype)
}
\arguments{
\item{x}{binmap object}

\item{outputfile}{output file name}

\item{filetype}{filetype}
}
\value{
result file
}
\description{
Choose output file type
}
