% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readgeno.R
\name{readgeno}
\alias{readgeno}
\title{Data Input}
\usage{
readgeno(
  inputfile,
  datatype,
  outputfile,
  screening = FALSE,
  maf = NA,
  geno = NA,
  mind = NA,
  hwe = NA,
  father = "1/1",
  mother = "0/0"
)
}
\arguments{
\item{inputfile}{input file name}

\item{datatype}{inputfile type. support vcf/ped/bed}

\item{outputfile}{output file name}

\item{screening}{a logical value. use plink to screening}

\item{maf}{if screening == T ,it useful}

\item{geno}{if screening == T ,it useful}

\item{mind}{if screening == T ,it useful}

\item{hwe}{if screening == T ,it useful}

\item{father}{support a character vector or a number that father index of sample}

\item{mother}{support a character vector or a number that father index of sample}
}
\value{
data.frame
}
\description{
Read a file in table format from vcf/ped/bed
}
